start initializing connections
