+++ lib/environment.cc
@@ -126,7 +126,7 @@ auto Environment::startup(const TimePoint& start_time) -> std::thread {
     }
   });
 }
-void Environment::draw_connection(std::size_t source, std::size_t sink, Environment::ConnectionProperties properties) {
+void Environment::draw_connection(std::size_t source, std::size_t sink, ConnectionProperties properties) {
   graph_.add_edge(source, sink, properties);
 }
